Implemented Level Parsing
The levels can now read from the given file path on construction and parse the data to create a level and spawn point.

On top of that I've unified it with the Texture Asset by creating a standard Asset class which is very basic and making both the texture class and the Level class inherit from it, this doesn't provide much functionality right now but gives me the ability to expand in future.

Next I'm gonna be working on setting up the level manager to be able to handle the levels once they are parsed and in the program's memory.