IMPRESSIVE DDEBUG
I'm so incredibly proud of myself for this- There was a memory leak when moving onto the next level right? Weird. Cuz I specifically made sure there was no  memory leaks when deleting a level. I had the weirdest hunch that it came from the Texture2D class for some reason, I was leaking megabytes at a time after all what else could it be? And I was right. I wasn't freeing the texture memory on the GPU when deleting the Texture class. I fixed that but then textures wouldn't load, I did a bit of debugging, checking the call stack all that fancy stuff and I realised it was because in the renderer I was passing in textures by clone parameters not by pointer so I was deleting the texture I wanted to render at the end of the scope. Silly. So I changed the renderer to only take pointers to textures to fix this issue. No more memory leaks, very cool.

Also changed the level manager to pick level order based off of a levelList text file in the levels directory, small change but its just nice, might neaten up the code tho.