Full* Debug Menu implimented
I say Full* and not just Full because I might add more to this menu as I want/need it in the future. That being said, as of now, the debug menu has all the features I want, from Level navigation, to player stats to performance stats, all in a nice, transparent, togglable (via F3) menu.